chore: move AbstractWebMessageResponse to api
